% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_phoenix_suns.R
\name{add_purple_gridlines}
\alias{add_purple_gridlines}
\title{A helper function that makes the gridlines purple}
\usage{
add_purple_gridlines(gridline_color = "#5350ea")
}
\arguments{
\item{gridline_color}{color of the gridlines}
}
\value{
a plot with purple gridlines
}
\description{
A helper function that makes the gridlines purple
}
