% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_midnight.R
\name{change_background_color}
\alias{change_background_color}
\title{Adds the background color to the midnight theme}
\usage{
change_background_color(background_color)
}
\arguments{
\item{background_color}{The color of the plot's background}
}
\value{
A ggplot theme with a background color
}
\description{
Adds the background color to the midnight theme
}
