% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_plots.R
\docType{data}
\name{dark_pink}
\alias{dark_pink}
\title{A function that will automatically plot one variable data with exotic colors}
\format{
An object of class \code{character} of length 1.
}
\usage{
dark_pink
}
\arguments{
\item{data}{dataset to apply colors to}

\item{x}{the x variable}

\item{y}{the y variable}

\item{adp_colors}{the color to apply to the visualization}

\item{type}{type of visualization}
}
\value{
the plot colored with specified color
}
\description{
A function that will automatically plot one variable data with exotic colors
}
\keyword{datasets}
