% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_midnight.R
\name{theme_midnight}
\alias{theme_midnight}
\title{Midnight Theme for ggplot graph}
\usage{
theme_midnight(
  text_color = "wheat",
  title_color = "wheat",
  background_color = "gray40",
  gridline_color = "black"
)
}
\arguments{
\item{text_color}{The color of the text and axis labels}

\item{title_color}{The color of the title, subtitle, and caption}

\item{background_color}{The color of the plot's background}

\item{gridline_color}{The color of the gridlines}
}
\value{
A ggplot midnight theme'
}
\description{
Midnight Theme for ggplot graph
}
