% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_phoenix_suns.R
\name{theme_phoenix_suns}
\alias{theme_phoenix_suns}
\title{A theme based on the colors of the Phoenix Suns}
\usage{
theme_phoenix_suns(
  text_color = "#E56020",
  title_color = "#E56020",
  panel_color = "#E56020",
  gridline_color = "#5350ea",
  background_color = "#5350ea",
  title_font = "sans",
  text_font = "sans"
)
}
\arguments{
\item{text_color}{color of the text}

\item{title_color}{color of the title}

\item{panel_color}{color of the panels of the plot}

\item{gridline_color}{color of the gridlines}

\item{background_color}{color of background of the plot}
}
\value{
a plot with a Phoenix Suns theme
}
\description{
A theme based on the colors of the Phoenix Suns
}
