% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ride_high_mustangs.R
\name{theme_ride_high_mustangs}
\alias{theme_ride_high_mustangs}
\title{A custom theme for ggplot2 visualizations based off of Cal Poly! Ride High Mustangs}
\usage{
theme_ride_high_mustangs(plot)
}
\arguments{
\item{plot}{plot that theme will be applied to}

}
\description{
A custom theme for ggplot2 visualizations based off of Cal Poly! Ride High Mustangs
}
