% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ride_high_mustangs.R
\name{try_theme_ride_high_mustangs}
\alias{try_theme_ride_high_mustangs}
\title{This is a helper function that allows users to try out the Cal Poly Theme!}
\usage{
try_theme_ride_high_mustangs(plot)
}
\arguments{
\item{plot}{Plot that you want to test the Cal Poly theme on}
}
\description{
This is a helper function that allows users to try out the Cal Poly custom Theme!
}
